Update dependencies and improve the build
- Updates
  - PyInstaller 3.6 -> 4.2
  - PyYAML 5.3.1 -> 5.4.1
  - Sphinx 2.4.4 -> 3.4.3
  - recommonmark 0.6.0 -> 0.7.1
  - sphinxcontrib-opanapi 0.6.0 -> 0.7.0
  - sphinxcontrib-plantuml 0.18 -> 0.19
    - Pillow 7.1.1 -> 8.1.0
  - sphinxcontrib-redoc 1.5.1 -> 1.6.0
  - sphinxcontrib-websupport 1.1.2 -> 1.2.4
  - sphinxcontrib-youtube
  - sphinxemoji 0.1.5 -> 0.1.8
  - sphinx-markdown-tables 0.0.12 -> 0.0.15
  - sphinx_bootstrap_theme 0.7.1 -> 0.8.0
  - sphinx_rtd_theme 0.4.3 -> 0.5.1
- Build improvements
  - Made it easier to generate a patch by keeping the original venv and
    the patched venv separately.
  - Fixed recommonmark usage in `test_site`
  - Updated PlantUML from 1.2019.7 to 1.2021.1 in `test_site`

+++ 3-apply-patches.sh
@@ -1,18 +1,33 @@
 #!/bin/bash -e
-export PATH="$PWD/build/venv/bin:$PWD/build/venv/Scripts:$PATH"
-if [[ ! -d "$PWD/build/venv" ]]; then
+ORIG_VENV_HOME="$PWD/build/venv.orig"
+NEW_VENV_HOME="$PWD/build/venv.new"
+
+if [[ ! -d "$ORIG_VENV_HOME" ]]; then
   echo "virtualenv not ready"
   exit 1
 fi
 
-# Apply some patches and recompile the patched files.
+# Revert the new venv so that we can patch from scratch.
+rsync -aiP --delete "$ORIG_VENV_HOME/" "$NEW_VENV_HOME/"
+
+# Make sure we don't refer to the old venv.
+grep -rFl "venv.orig" "$NEW_VENV_HOME/bin" | while read F; do
+  perl -pi -e "s#venv.orig#venv.new#g" "$F"
+done
+if grep -rFl "venv.orig" "$NEW_VENV_HOME"; then
+  echo "The new virtualenv contains a reference to the old one."
+  exit 1
+fi
+
+# Apply some patches.
+export PATH="$NEW_VENV_HOME/bin:$NEW_VENV_HOME/Scripts:$PATH"
 PATCH_DIR="$PWD/patches"
-if [[ -d build/venv/Lib/site-packages ]]; then
-  SITEPKG_DIR="build/venv/Lib/site-packages"
-elif [[ -d build/venv/lib/python3.7/site-packages ]]; then
-  SITEPKG_DIR="build/venv/lib/python3.7/site-packages"
+if [[ -d "$NEW_VENV_HOME/Lib/site-packages" ]]; then
+  SITEPKG_DIR="$NEW_VENV_HOME/Lib/site-packages"
+elif [[ -d "$NEW_VENV_HOME/lib/python3.7/site-packages" ]]; then
+  SITEPKG_DIR="$NEW_VENV_HOME/lib/python3.7/site-packages"
 else
-  SITEPKG_DIR="build/venv/lib/python3.6/site-packages"
+  SITEPKG_DIR="$NEW_VENV_HOME/lib/python3.6/site-packages"
 fi
 pushd "$SITEPKG_DIR"
 if [[ ! -a sphinxcontrib/__init__.py ]]; then
@@ -21,6 +36,11 @@ fi
 find "$PATCH_DIR" -name '*.diff' -print | while read -r PATCH; do
   patch -p1 -i "$PATCH"
 done
-find . -type d -name '__pycache__' -exec rm -fr {} ';' >/dev/null 2>&1 || true
-python -m compileall . >/dev/null 2>&1
+find . -type f -name '*.orig' -delete || true
+
+# Pre-compile everything unless 'nocompile' option is specified.
+if [[ "$1" != 'nocompile' ]]; then
+  find . -type d -name '__pycache__' -exec rm -fr {} ';' >/dev/null 2>&1 || true
+  python -m compileall . >/dev/null 2>&1
+fi
 popd
